Crate nargo_fmt

The Noir formatter.

It works by using two techniques:

Surfing a parsed module by using a lexer.

Trying to not exceed the maximum line width by formatting to intermediate chunk (see chunks.rs)

What is lexer surfing?

Suppose we need to format this code:

fn foo ( ) { let x : Field = ( 2 , 3 ) ; }

We first parse the above code so we end up with a ParsedModule. Next we traverse this module contents and process each item, statement, expression, type, etc., we find.

For example, the first thing we’ll find is a function. We know it has no visibility and no doc comments so we can expect an fn keyword to be there. We write it. Next will come the identifier. For that we “skip” any spaces and comments between fn and foo, writing only one space instead of possibly multiple spaces, then write “foo”. If there were comments between fn and foo, we write them (the formatter will never lose comments).

Next we know there are no generics, so we can expect a (, etc.

In this way we go token by token, inserting newlines when needed, removing extra spaces, indenting things as we go deep inside structures, etc.

But that’s not all. The formatter will try to not exceed the configurable maximum width. It will do that but, for simplicity, only for function parameters list, statements and expressions (we assume an impl Foo ... line won’t exceed the maximum length, and if it does it’s not a big deal, or we can always improve things later). For this, read the comments in chunks.rs.

Modules

chunks 🔒

This module has all the logic to format a series of chunks (a piece of text) in a way that we (almost always) never exceed the configurable maximum line width.

config 🔒

errors

formatter 🔒

Structs

Config

Enums

ImportsGranularity

How imports should be grouped into use statements. Imports will be merged or split to the configured level of granularity.

Functions

format