nargo_fmt::chunks
Struct ChunkFormatter

pub(crate) struct ChunkFormatter<'a, 'b>(&'b mut Formatter<'a>);

Interface for creating TextChunks.

Tuple Fields
0: &'b mut Formatter<'a>
Implementations

impl<'a, 'b> ChunkFormatter<'a, 'b>

pub(crate) fn new(formatter: &'b mut Formatter<'a>) -> Self

pub(crate) fn chunk(&mut self, f: impl FnOnce(&mut Formatter<'_>)) -> TextChunk

Stops writing to the current buffer for the duration of the f call, which takes a formatter to write to. Then, returns a TextChunk with the written text.

This allows a caller to format pieces of code and then pre-process them before writing to the main buffer.

pub(crate) fn skip_comments_and_whitespace_chunk(&mut self) -> TextChunk

Stops writing to the current buffer, skips comments and whitespaces (formatting them) and returns the formatted result as a TextChunk.

pub(crate) fn new_group_tag(&mut self) -> GroupTag

pub(crate) fn bump(&mut self) -> Token

impl ChunkFormatter<'_, '_>

pub(super) fn format_expression( &mut self, expression: Expression, group: &mut ChunkGroup, )

fn format_literal(&mut self, literal: Literal, group: &mut ChunkGroup)

fn format_array_literal( &mut self, literal: ArrayLiteral, is_slice: bool, ) -> ChunkGroup

fn format_tuple(&mut self, exprs: Vec<Expression>) -> ChunkGroup

fn format_lambda(&mut self, lambda: Lambda) -> FormattedLambda

fn format_parenthesized_expression(&mut self, expr: Expression) -> ChunkGroup

pub(super) fn format_quote(&mut self) -> ChunkGroup

pub(super) fn format_comptime_expression( &mut self, block: BlockExpression, force_multiple_lines: bool, ) -> ChunkGroup

pub(super) fn format_unsafe_expression( &mut self, block: BlockExpression, force_multiple_lines: bool, ) -> ChunkGroup

pub(super) fn format_type_path(&mut self, type_path: TypePath) -> ChunkGroup

pub(super) fn format_expressions_separated_by_comma( &mut self, exprs: Vec<Expression>, force_trailing_comma: bool, group: &mut ChunkGroup, ) -> usize

Returns the maximum width of each expression to format. For example, if the list is [1, 234, 56], the maximum width is 3 (that of 234).

pub(super) fn format_items_separated_by_comma<Item, F>( &mut self, items: Vec<Item>, force_trailing_comma: bool, surround_with_spaces: bool, group: &mut ChunkGroup, format_item: F, )
where F: FnMut(&mut Self, Item, &mut ChunkGroup),

fn format_constructor( &mut self, constructor: ConstructorExpression, ) -> ChunkGroup

fn format_member_access( &mut self, member_access: MemberAccessExpression, ) -> ChunkGroup

fn format_member_access_impl( &mut self, member_access: MemberAccessExpression, nested: bool, group_tag: GroupTag, ) -> ChunkGroup

fn format_cast(&mut self, cast_expression: CastExpression) -> ChunkGroup

fn format_prefix(&mut self, prefix: PrefixExpression) -> ChunkGroup

fn format_infix_expression(&mut self, infix: InfixExpression) -> ChunkGroup

fn format_infix_expression_with_group_tag( &mut self, infix: InfixExpression, group_tag: GroupTag, nested: bool, ) -> ChunkGroup

pub(super) fn format_if_expression( &mut self, if_expression: IfExpression, force_multiple_lines: bool, ) -> ChunkGroup

pub(super) fn format_if_expression_with_group_tag( &mut self, if_expression: IfExpression, force_multiple_lines: &mut bool, group_tag: GroupTag, ) -> ChunkGroup

pub(super) fn format_match_expression( &mut self, match_expression: MatchExpression, ) -> ChunkGroup

pub(super) fn format_match_expression_with_group_tag( &mut self, match_expression: MatchExpression, group_tag: GroupTag, ) -> ChunkGroup

fn format_index_expression(&mut self, index: IndexExpression) -> ChunkGroup

fn format_call(&mut self, call: CallExpression) -> ChunkGroup

fn format_method_call( &mut self, method_call: MethodCallExpression, ) -> ChunkGroup

fn format_method_call_impl( &mut self, method_call: MethodCallExpression, nested: bool, group_tag: GroupTag, ) -> ChunkGroup

fn format_constrain( &mut self, constrain_statement: ConstrainExpression, ) -> ChunkGroup

pub(super) fn format_block_expression( &mut self, block: BlockExpression, force_multiple_lines: bool, ) -> ChunkGroup

pub(super) fn format_block_expression_contents( &mut self, block: BlockExpression, force_multiple_lines: bool, group: &mut ChunkGroup, )

pub(super) fn format_non_empty_block_expression_contents( &mut self, block: BlockExpression, force_multiple_lines: bool, group: &mut ChunkGroup, )

pub(super) fn empty_block_contents_chunk(&mut self) -> Option<ChunkGroup>

impl ChunkFormatter<'_, '_>

pub(super) fn format_global( &mut self, let_statement: LetStatement, visibility: ItemVisibility, ) -> ChunkGroup

impl ChunkFormatter<'_, '_>

pub(super) fn format_pattern(&mut self, pattern: Pattern) -> ChunkGroup

impl ChunkFormatter<'_, '_>

pub(super) fn format_statement( &mut self, statement: Statement, group: &mut ChunkGroup, ignore_next: bool, )

fn format_let_statement(&mut self, let_statement: LetStatement) -> ChunkGroup

pub(super) fn format_let_or_global( &mut self, keyword: Keyword, pattern: Pattern, typ: Option<UnresolvedType>, value: Option<Expression>, attributes: Vec<SecondaryAttribute>, ) -> ChunkGroup

fn format_assign(&mut self, assign_statement: AssignStatement) -> ChunkGroup

fn format_for_loop(&mut self, for_loop: ForLoopStatement) -> ChunkGroup

fn format_loop(&mut self, loop_: LoopStatement) -> ChunkGroup

fn format_while(&mut self, while_: WhileStatement) -> ChunkGroup

fn format_comptime_statement(&mut self, statement: Statement) -> ChunkGroup

fn format_semi_statement(&mut self, expression: Expression) -> ChunkGroup

impl ChunkFormatter<'_, '_>

pub(super) fn format_import( &mut self, use_tree: UseTree, visibility: ItemVisibility, ) -> ChunkGroup

fn format_use_tree(&mut self, use_tree: UseTree) -> ChunkGroup