nargo_fmt::chunks
Struct ChunkGroup

pub(crate) struct ChunkGroup {
    pub(crate) chunks: Vec<Chunk>,
    pub(crate) one_chunk_per_line: bool,
    pub(crate) force_multiple_lines: bool,
    pub(crate) tag: Option<GroupTag>,
    pub(crate) kind: GroupKind,
    pub(crate) force_multiline_on_children_with_same_tag_if_multiline: bool,
}

Fields
chunks: Vec<Chunk>one_chunk_per_line: bool

If true, when formatting in multiple lines, and after a SpaceOrLine, a line will be written. If false, when formatting in multiple lines, and after a SpaceOrLine, a space will be inserted and the next chunk will go in the same line if it fits that line.

This is used to, for example, control how arrays are formatted. If each element is short, we’ll format the array like this:

[ 1, 2, 3, 4, 5 ]

but if one of the elements is long, each one will go in a separate line:

[
    1,
    1234567890123,
    3
]

force_multiple_lines: bool

If true, regardless of this group’s chunks, this group will be formatted in multiple lines. This is set to true when, for example, we format a block that has at least two statements: we always want to show that in multiple lines.

tag: Option<GroupTag>

Groups can be tagged. For example we tag all consequences and alternative blocks of an if expression. If we determine one of them needs to be formatted in multiple lines, we find all other chunks with the same tag and mark them too to be formatted in multiple lines.

kind: GroupKind

The kind of this group. Some group kinds are formatted in a special way (mainly lambda arguments that are the last expression in a list).

force_multiline_on_children_with_same_tag_if_multiline: bool

This name is a bit long and explicit, but it’s to make things clearer: if we determine that this group needs to be formatted in multiple lines, children groups with the same tag will also be formatted in multiple lines.

This is used for example in infix expressions like a + b + c + d, where if we determine that a + b needs to be formatted in multiple lines, we want the entire tree (of those infix expressions) to be formatted in multiple lines.

Implementations

impl ChunkGroup

pub(crate) fn new() -> Self

pub(crate) fn text(&mut self, chunk: TextChunk)

Appends a text to this group. If the last chunk in this group is a text or verbatim, no new chunk is inserted and instead the last text chunk is extended.

pub(crate) fn verbatim(&mut self, chunk: TextChunk)

Appends a verbatim text chunk to this group.

pub(crate) fn text_impl(&mut self, chunk: TextChunk, verbatim: bool)

pub(crate) fn space(&mut self, formatter: &mut ChunkFormatter<'_, '_>)

Appends a single space to this group by reading it from the given formatter.

pub(crate) fn semicolon(&mut self, formatter: &mut ChunkFormatter<'_, '_>)

Appends a semicolon to this group by reading it from the given formatter. This will actually end up attaching the semicolon to the last text in this group so that we don’t end up with stray semicolons.

pub(crate) fn text_attached_to_last_group(&mut self, chunk: TextChunk)

Appends a TextChunk to this chunks chunks. However, if the last chunk is a group, it’s appended to that group’s last text.

pub(crate) fn trailing_comment(&mut self, chunk: TextChunk)

Appends a trailing comment (it’s formatted slightly differently than a regular text chunk).

pub(crate) fn leading_comment(&mut self, chunk: TextChunk)

Appends a leading comment (it’s formatted slightly differently than a regular text chunk).

pub(crate) fn trailing_comma(&mut self)

Appends a trailing comma (will only show up if the group is formatted in multiple lines).

pub(crate) fn group(&mut self, group: ChunkGroup)

Appends another group as a nested group.

pub(crate) fn line(&mut self)

Append one line to this chunk.

pub(crate) fn lines(&mut self, two: bool)

Append one or two lines to this chunk.

pub(crate) fn space_or_line(&mut self)

Appends a SpaceOrLine chunk, which means that it’s a space when this group is formatted in a single line, or a line when it’s formatted in multiple lines.

pub(crate) fn increase_indentation(&mut self)

pub(crate) fn decrease_indentation(&mut self)

pub(crate) fn push_indentation(&mut self)

pub(crate) fn pop_indentation(&mut self)

pub(crate) fn push(&mut self, chunk: Chunk)

pub(crate) fn width(&self) -> usize

pub(crate) fn expression_list_width(&self) -> usize

pub(crate) fn has_newlines(&self) -> bool

pub(crate) fn has_lambda_as_last_expression_in_list(&self) -> bool

Determines if this group has a LambdaAsLastExpressionInList chunk. Note that if this group is a MethodCall, this is checked for the ExpressionList group inside it.

pub(crate) fn set_lambda_as_last_expression_in_list_indentation( &mut self, indentation_to_set: i32, )

Finds the LambdaAsLastExpressionInList associated to this group and sets its indentation to the given value.

pub(crate) fn prepare_for_multiple_lines(self) -> ChunkGroup

Before writing a Chunks object in multiple lines, create a new one where TrailingComma is turned into Text. Because Chunks will glue two consecutive Texts together, if we have two chunks Text("123"), TrailingComma, we’ll consider the entire string “123,” when deciding whether we can still write in the current line or not.

fn add_trailing_comma_to_last_text(&mut self)

fn width_until_line(&self) -> (usize, bool)

Returns the width of text until we hit a Line or LineOrSpace, together with whether we hit a Line or LineOrSpace.

fn first_group(&self) -> Option<&ChunkGroup>

fn has_expression_list_or_method_call_group(&self) -> bool

fn method_call_lambda_first_line_width(&self) -> Option<usize>

Assuming this is a MethodCall group, if the ExpressionList nested in it has a LambdaAsLastExpressionInList, returns its first_line_width.