nargo_fmt::formatter
Struct Formatter

pub(crate) struct Formatter<'a> {
    pub(crate) config: &'a Config,
    source: &'a str,
    lexer: Lexer<'a>,
    token: Token,
    token_span: Span,
    pub(crate) indentation: i32,
    indentation_stack: Vec<i32>,
    written_comments_count: usize,
    ignore_next: bool,
    pub(crate) group_tag_counter: usize,
    pub(crate) max_width: usize,
    pub(crate) buffer: Buffer,
}

Fields
config: &'a Configsource: &'a strlexer: Lexer<'a>token: Tokentoken_span: Spanindentation: i32

The current indentation level. We allow it to be negative because in some cases we just want to decrease indentation to preemptively cancel out an indentation that will come later which we don’t want to take effect, and we don’t want to panic when reaching those negative values.

indentation_stack: Vec<i32>

When formatting chunks we sometimes need to remember the current indentation and restore it later. This is what this stack is used for.

written_comments_count: usize

Whenever a comment is written, this counter is incremented. In this way we can know if comments were written while formatting some code: we remember the previous value, format, then see if it increased. This is used, for example, when transforming foo::{bar} into foo::bar: we only do that if there were no comments between { and }.

ignore_next: bool

If we find a comment like this one:

// noir-fmt:ignore

we won’t format the next node (in some cases: only applies to statements and items).

group_tag_counter: usize

A counter to create GroupTags.

max_width: usize

We keep a copy of the config’s max width because when we format chunk groups we somethings change this so that a group has less space to write to.

buffer: Buffer

This is the buffer where we write the formatted code.

Implementations

impl<'a> Formatter<'a>

pub(crate) fn chunk_formatter(&mut self) -> ChunkFormatter<'a, '_>

Returns an object that has a chunk method to get a TextChunk. This method exists so that we can’t mix the two operation modes: using the formatter directly while writing to the buffer, or creating text chunks.

pub(crate) fn format_chunk_group(&mut self, group: ChunkGroup)

Main interface to format a chunk group. Here it’s determined if the chunk will group in a single line or multiple lines.

pub(crate) fn format_chunk_group_impl(&mut self, group: ChunkGroup)

pub(crate) fn format_chunks_group_impl_without_lambda_handling( &mut self, group: ChunkGroup, )

pub(crate) fn format_chunk_group_in_one_line(&mut self, group: ChunkGroup)

pub(crate) fn format_chunk_group_in_multiple_lines(&mut self, group: ChunkGroup)

fn format_lambda_body_removing_braces(&mut self, group: ChunkGroup)

fn write_chunk_lines(&mut self, string: &str)

Appends the string to the current buffer line by line, with some pre-checks.

pub(crate) fn new_group_tag(&mut self) -> GroupTag

Returns a new GroupTag that is unique compared to other new_group_tag calls.

impl Formatter<'_>

pub(super) fn format_type_alias(&mut self, type_alias: TypeAlias)

impl Formatter<'_>

pub(super) fn format_attributes(&mut self, attributes: Attributes)

pub(super) fn format_secondary_attributes( &mut self, attributes: Vec<SecondaryAttribute>, )

fn format_attribute(&mut self, attribute: Attribute)

fn format_function_attribute(&mut self, attribute: FunctionAttribute)

pub(super) fn format_secondary_attribute( &mut self, attribute: SecondaryAttribute, )

fn format_deprecated_attribute(&mut self, message: Option<String>)

fn format_test_attribute(&mut self, test_scope: TestScope)

fn format_fuzz_attribute(&mut self, fuzz_scope: FuzzingScope)

fn format_meta_attribute(&mut self, meta_attribute: MetaAttribute)

fn format_must_use_attribute(&mut self, message: Option<String>)

fn format_no_args_attribute(&mut self)

fn format_one_arg_attribute(&mut self)

fn format_tag_attribute(&mut self, span: Span)

Removes the trailing whitespaces from the tag attribute.

impl Formatter<'_>

pub(crate) fn write_space(&mut self)

Writes a single space, skipping any whitespace and comments. That is, suppose the next token is a big whitespace, possibly with multiple lines. Those are skipped but only one space is written. In this way if we have “mod foo” it’s transformed to “mod foo”. If there are comments in between mod and foo they are written, though! No comment is ever lost.

A space is not appended to the buffer is it already ends with a space.

pub(crate) fn write_space_without_skipping_whitespace_and_comments(&mut self)

Writes a single space, but doesn’t skip whitespace and comments before doing that.

A space is not appended to the buffer is it already ends with a space.

pub(crate) fn skip_whitespace(&mut self)

pub(crate) fn skip_whitespace_if_it_is_not_a_newline(&mut self)

Only skips whitespace if it doesn’t have newlines in it. Note that this doesn’t write whitespace or comments at all.

pub(crate) fn skip_comments_and_whitespace(&mut self)

Skips comments and whitespace, writing newlines if there are any. If there are multiple consecutive newlines, only one is written.

pub(crate) fn skip_comments_and_whitespace_writing_multiple_lines_if_found( &mut self, )

Similar to skip_comments_and_whitespace, but will write two lines if multiple newlines are found (but at most two lines at a time).

pub(crate) fn skip_comments_and_whitespace_impl( &mut self, write_multiple_lines: bool, at_beginning: bool, )

pub(crate) fn following_newlines_count(&self) -> usize

Returns the number of newlines that come next, if we are at a whitespace token (otherwise returns 0).

pub(crate) fn write_line(&mut self)

Writes a single newline, if the last thing we wrote wasn’t also a newline (this prevents multiple consecutive newlines, though that’s still possible to do if you call write_multiple_lines_...).

Any whitespace or comments found right at and after the current token are “skipped” (whitespace is discarded, comments are written).

pub(crate) fn write_line_without_skipping_whitespace_and_comments( &mut self, ) -> bool

pub(crate) fn write_multiple_lines_without_skipping_whitespace_and_comments( &mut self, )

pub(crate) fn trim_spaces(&mut self)

Trim spaces from the end of the buffer.

pub(crate) fn trim_comma(&mut self) -> bool

Trim commas from the end of the buffer. Returns true if a comma was trimmed.

impl Formatter<'_>

pub(super) fn format_inner_doc_comments(&mut self)

pub(super) fn format_outer_doc_comments(&mut self)

pub(super) fn format_outer_doc_comments_checking_safety(&mut self)

Formats outer doc comments, turning them into regular comments if they start with “Safety:”

impl Formatter<'_>

pub(super) fn format_enum(&mut self, noir_enum: NoirEnumeration)

impl Formatter<'_>

pub(super) fn format_empty_block_contents(&mut self)

impl Formatter<'_>

pub(super) fn format_function( &mut self, func: NoirFunction, skip_visibility: bool, )

pub(super) fn format_function_impl(&mut self, func: FunctionToFormat)

pub(super) fn format_function_modifiers( &mut self, visibility: ItemVisibility, skip_visibility: bool, )

pub(super) fn format_function_parameters( &mut self, parameters: Vec<Param>, group: &mut ChunkGroup, )

fn format_function_param(&mut self, param: Param)

pub(super) fn format_function_right_paren_until_left_brace_or_semicolon( &mut self, return_type: FunctionReturnType, visibility: Visibility, has_where_clause: bool, has_parameters: bool, semicolon: bool, group: ChunkGroup, )

Returns whether the left brace of semicolon was written (we don’t write it when there’s a comment before those tokens)

fn skip_stray_where_keyword(&mut self)

fn format_function_return_type( &mut self, return_type: FunctionReturnType, visibility: Visibility, )

pub(super) fn format_function_body(&mut self, body: BlockExpression)

impl Formatter<'_>

pub(super) fn format_generics(&mut self, generics: Vec<UnresolvedGeneric>)

fn format_generic(&mut self, generic: UnresolvedGeneric)

pub(super) fn format_generic_type_args(&mut self, generics: GenericTypeArgs)

impl Formatter<'_>

pub(super) fn format_global( &mut self, let_statement: LetStatement, visibility: ItemVisibility, )

impl Formatter<'_>

pub(super) fn format_impl(&mut self, type_impl: TypeImpl)

impl Formatter<'_>

pub(super) fn format_items(&mut self, items: Vec<Item>, ignore_next: bool)

pub(super) fn format_item(&mut self, item: Item, ignore_next: bool)

fn next_import_group(&self, items: &mut Vec<Item>) -> Option<ImportGroup>

Returns the next import group, if there’s is one.

An import group is one or more use statements that all have the same visibility, as long as exactly one newline separates them, and as long as there are no comments in the use statements or trailing comments in them.

Each import group will be sorted and merged, if the configuration is set to do so.

fn span_has_comments(&self, span: Span) -> bool

fn span_starts_with_trailing_comment(&self, span: Span) -> bool

fn span_is_import_group_separator(&self, span: Span) -> bool

Returns true if there at most one newline in the given span and it contains no comments.

impl Formatter<'_>

pub(super) fn format_lvalue(&mut self, lvalue: LValue)

impl Formatter<'_>

pub(super) fn format_module_declaration( &mut self, module_declaration: ModuleDeclaration, )

pub(super) fn format_submodule(&mut self, submodule: ParsedSubModule)

impl Formatter<'_>

pub(super) fn format_path(&mut self, path: Path)

pub(super) fn format_turbofish(&mut self, generics: Vec<UnresolvedType>)

impl Formatter<'_>

pub(super) fn format_pattern(&mut self, pattern: Pattern) -> ChunkGroup

impl Formatter<'_>

pub(super) fn format_struct(&mut self, noir_struct: NoirStruct)

impl Formatter<'_>

pub(super) fn format_trait_impl(&mut self, trait_impl: NoirTraitImpl)

fn format_trait_impl_item(&mut self, item: TraitImplItem)

impl Formatter<'_>

pub(super) fn format_trait(&mut self, noir_trait: NoirTrait)

fn format_trait_item(&mut self, item: TraitItem)

impl Formatter<'_>

pub(super) fn format_type_expression( &mut self, type_expr: UnresolvedTypeExpression, )

impl Formatter<'_>

pub(super) fn format_type(&mut self, typ: UnresolvedType)

pub(super) fn format_as_trait_path(&mut self, as_trait_path: AsTraitPath)

impl Formatter<'_>

pub(super) fn format_import( &mut self, use_tree: UseTree, visibility: ItemVisibility, )

impl Formatter<'_>

pub(super) fn merge_and_format_imports( &mut self, imports: Vec<UseTree>, visibility: ItemVisibility, )

fn format_import_tree( &mut self, import_tree: ImportTree, visibility: ItemVisibility, )

impl Formatter<'_>

pub(super) fn format_item_visibility(&mut self, visibility: ItemVisibility)

pub(super) fn format_visibility(&mut self, visibility: Visibility)

impl Formatter<'_>

pub(super) fn format_where_clause( &mut self, constraints: Vec<UnresolvedTraitConstraint>, write_trailing_comma_and_new_line: bool, )

pub(super) fn format_trait_bounds(&mut self, trait_bounds: Vec<TraitBound>)

pub(super) fn format_trait_bound(&mut self, trait_bound: TraitBound)

impl<'a> Formatter<'a>

pub(crate) fn new(source: &'a str, config: &'a Config) -> Self

pub(crate) fn format_program(&mut self, parsed_module: ParsedModule)

pub(crate) fn format_parsed_module( &mut self, parsed_module: ParsedModule, ignore_next: bool, )

pub(crate) fn write_identifier(&mut self, ident: Ident)

pub(crate) fn write_identifier_or_integer(&mut self, ident: Ident)

pub(crate) fn write_left_paren(&mut self)

pub(crate) fn write_right_paren(&mut self)

pub(crate) fn write_left_brace(&mut self)

pub(crate) fn write_right_brace(&mut self)

pub(crate) fn write_left_bracket(&mut self)

pub(crate) fn write_right_bracket(&mut self)

pub(crate) fn write_comma(&mut self)

pub(crate) fn write_semicolon(&mut self)

pub(crate) fn write_keyword(&mut self, keyword: Keyword)

Writes the given keyword, if the current token is that keyword (so this is a check that we are producing a token we expect to be in the source we are traversing). Then advances to the next token.

Calls write_token so comments and whitespaces are skipped before writing the keyword.

pub(crate) fn write_token(&mut self, token: Token)

Writes the given token, if the current token is the same as the given one (so this is a check that we are producing a token we expect to be in the source we are traversing). Then advances to the next token.

Before writing the token any comments and spaces are skipped. This is so that a caller can call write_token, write_keyword, write_space, etc., without having to explicitly call skip_comments_and_whitespace in between those calls.

pub(crate) fn write_current_token(&mut self)

Writes the current token but doesn’t advance to the next one.

pub(crate) fn write_current_token_and_bump(&mut self)

Writes the current token and advances to the next one

pub(crate) fn write_current_token_trimming_end(&mut self)

Writes the current token trimming its end but doesn’t advance to the next one. Mainly used when writing comment lines, because we never want trailing spaces inside comments.

pub(crate) fn write_current_token_as_in_source(&mut self)

Writes the current token but without turning it into a string using to_string(). Instead, we check the token’s span and format what’s in the original source there (useful when formatting integer tokens, because a token like 0xFF ends up being an integer with a value 255, but we don’t want to change 0xFF to 255).

pub(crate) fn write_source_span(&mut self, span: Span)

Writes whatever is in the given span relative to the file’s source that’s being formatted.

pub(crate) fn write_source_span_trimmed(&mut self, span: Span)

Writes whatever is in the given span relative to the file’s source that’s being formatted but trims the whitespaces at the end.

pub(crate) fn write_indentation(&mut self)

Writes the current indentation to the buffer, but only if the buffer is empty or it ends with a newline (otherwise we’d be indenting when not needed).

pub(crate) fn write_and_skip_span_without_formatting(&mut self, span: Span)

Writes whatever is in the source at the given span without formatting it, then advances the lexer until past the end of the span. This is mainly used to avoid formatting items and statements when a noir-fmt:ignore comment is found.

pub(crate) fn skip_past_span_end_without_formatting(&mut self, span_end: u32)

Advances the lexer until past the given span end without writing anything to the buffer.

pub(crate) fn write(&mut self, str: &str)

Writes a string to the buffer.

pub(crate) fn current_line_width(&self) -> usize

pub(crate) fn increase_indentation(&mut self)

pub(crate) fn decrease_indentation(&mut self)

pub(crate) fn push_indentation(&mut self)

pub(crate) fn pop_indentation(&mut self)

pub(crate) fn is_at_keyword(&self, keyword: Keyword) -> bool

pub(crate) fn is_at(&self, token: Token) -> bool

pub(crate) fn bump(&mut self) -> Token

Advances to the next token (the current token is not written).

pub(crate) fn read_token_internal(&mut self) -> SpannedToken