nargo_fmt::formatter
Module use_tree_merge

Structs

ImportTree 🔒

An import tree to represent merged imports. For example for the given imports:

Enums

Segment 🔒

We keep Crate, Super and Dep as special segments so that they are ordered in that way (they’ll come in that order before any plain segment).

Functions

format_merged_import 🔒

format_merged_import_with_visibility 🔒

merge_imports 🔒

Combines all use trees to form a single ImportTree.

merge_imports_in_tree 🔒