nargo_fmt
Module chunks

This module has all the logic to format a series of chunks (a piece of text) in a way that we (almost always) never exceed the configurable maximum line width.

It’s heavily inspired by this excellent blog post:

https://yorickpeterse.com/articles/how-to-write-a-code-formatter/

However, some changes were introduces to handle comments and other particularities of Noir.

Structs

ChunkFormatter 🔒

Interface for creating TextChunks.

ChunkGroup 🔒

GroupTag 🔒

TextChunk 🔒

A text chunk. It precomputes the text width and whether it has newlines.

Enums

Chunk 🔒

A chunk can either be text or a directive that instructs the formatter to do something (for example: increase or decrease the current indentation).

GroupKind 🔒